EquilibriumUtilities.jl Documentation

A small package of utilities commonly used to compute economic models.

Exports

"EquilibriumUtilities, WrappedDict, chunk, converge, diag_view, infnorm_pctdev, issquare, newton, normalise!, offdiag_view, quietly, update!, zero_safe"

Iterative convergence

The generalised iterative routine and its parametres:

EquilibriumUtilities.ConvergeParameters — Type

ConvergeParameters{T1 <: Real, T2 <: Integer, T3}

Contains optional parameters for the converge solver.

See also converge.

Fields

diff_tol::T1 = 1e-6: If the value returned by step_diff is less than this tolerance, stop.

up_tol::T1 = zero(diff_tol): If the value returned by update is less than this tolerance, stop.

max_iter::T2 = 200: Maximum number of iterations before giving up.

msg::T3 = "No convergence": Warning message to display if there isn't convergence within the maximum number of iterations.

verbose::Bool = false: Print the diff every iteration?

EquilibriumUtilities.converge — Method

converge(update::Function, step_diff::Function, init::Function; kw...) -> (converged, stalled_update)

Iterate until convergence. In particular, the problem is initiated with init(). Then, repeatedly:

apply step_diff() which should do an iteration step, then return the deviation

apply update(), which should return the size of the update

until convergence. The convergence criterion is controlled by keyword arguments; see ConvergeParameters.

The function returns a pair of booleans, the first of which indicates whether convergence was reached; the second of which indicates whether iteration was aborted because the update was too small.

For an example, see this package's tests which uses this function to solve the finite-firm CES game.

See also update!.

EquilibriumUtilities.converge — Method

converge(update::Function, step_diff::Function, init::Function, p::ConvergeParameters)

EquilibriumUtilities.infnorm_pctdev — Method

infnorm_pctdev(v1, v2)

Calculates the difference between two vectors by finding the maximum absolute percentage deviation.

EquilibriumUtilities.update! — Function

Optional helper function to use when writing an update() function for converge.

Update main according to secondary, with a dampening factor. Useful for iterative algorithms. Once complete, main will hold the updated value and secondary will hold main's original value (to keep a record of previous iteration).

Returns the norm of the update step (useful for breaking iteration if the step size is too small).

See also converge.

EquilibriumUtilities.update! — Method

update!(main, secondary; norm = infnorm_pctdev, dampen = 0.5)

See infnorm_pctdev.

Reactive dampening

Optional helpers, mostly implementing a dampening factor that dynamically adjusts as the equilibrium progresses.

The convergence routine above need not use this routine: the user can always define the update() function however they would like. However, this optional functionality allows the solver to choose a dampening factor taking into account the convergence path. It will be provided in the ds::DampenState struct so the user can access it when defining their update() function.

EquilibriumUtilities.DampenParameters — Type

struct DampenParameters{T1 <: Real, T2 <: Integer}

Parameters for dynamicdampen!.

Fields

loosen::T1 = -0.01: by how much to loosen the dampening factor

tighten::T1 = 0.01: by how much to tighten the dampening factor

min_dampen::T1 = 0.0: the minimum value for the dampen factor

max_dampen::T1 = 0.999: the maximum value for the dampen factor

scale::T1 = 0.925: the scale mentioned in dynamicdampen!

overshooting_share::T1 = 0.5: the share for isovershooting

grace_period::T2 = 50: how long to wait between adjustments

tighten_wait::T2 = 30: how long to wait between tightenings

loosen_wait::T2 = 40: how long to wait between loosenings

EquilibriumUtilities.DampenState — Type

struct DampenState{T <: AbstractVector{<: Real}}

Fields

dampenstats::T = [0.85; Inf; Inf; Inf]: contains, in order,

the dampening factor

number of iterations since the dampening factor lowered

number of iterations since the dampening factor increased

the diff used when checking whether iteration is far along enough to attempt loosening again

last_deviations::T2 = Float64[] for storing deviations from the last iteration

penultimate_deviations::T2 = Float64[] for storing deviations from the penultimate iteration

history::T2 = Float64[] for storing the history of diffs

See also dynamicdampen!, dampenfactor, dampenfactor!.

EquilibriumUtilities.converge — Method

converge(update::Function, step_diff::Function, init::Function, ds::DampenState, p::ConvergeParameters)

Wrapper that push!es the calculated diff to ds.history, so that ds is up-to-date.

EquilibriumUtilities.dampenfactor! — Method

dampenfactor!(ds::DampenState, factor)

Set the dampen factor in ds.

See also dampenfactor, DampenState.

EquilibriumUtilities.dampenfactor — Method

dampenfactor(ds::DampenState)

Retrieve the dampen factor from ds.

See also dampenfactor!, DampenState.

EquilibriumUtilities.dynamicdampen! — Method

dynamicdampen!(ds::DampenState, p::DampenParameters)

EquilibriumUtilities.dynamicdampen! — Method

dynamicdampen!(ds::DampenState; kw...)

Update the dampening factor based on convergence path. Strategy:

increase by ds.tighten if ds.history isdiverging or isovershooting

make no change if any of:

within the first p.grace_period iterations

it has been under p.tighten_wait iterations since the last tightening

it has been under p.loosen_wait iterations since the last loosening

the current difference is above p.scale*ds.reference_diff

otherwise, decrease by loosen

Fine tune control through keyword arguments; see DampenParameters.

See also DampenState, DampenParameters.

EquilibriumUtilities.isdiverging — Method

isdiverging(history)

Checks if iteration is on a bad path, by seeing if two of the past three iterations have worsened the difference. If updating is too aggressive, usually one of two things happens:

the difference blows up (if updating is too aggressive)

the difference oscillates between improving and worsening (if updating is a little too aggressive)

See also dynamicdampen!.

EquilibriumUtilities.isovershooting — Method

isovershooting(last_deviations, penultimate_deviations; share = 0.5)

Check if the convergence is overshooting each guess, as measured by a share of the deviations flipping signs between the last two iterations. If dampening isn't strong enough, the deviations alternate between positive and negative, wasting time by bouncing back and forth. Increasing the dampening can drastically speed up convergence by preventing this oscillation.

See also dynamicdampen!.

EquilibriumUtilities.update! — Method

update!(main, secondary, ds::DampenState; kw...)

Update using a dynamically-chosen dampening factor. However, if the user supplies the keyword argument dampen, it takes precedence over dynamically-chosen dampening factor.

See also dynamicdampen!, DampenState.

Newton method

EquilibriumUtilities.NewtonParameters — Type

NewtonParameters

Contains optional parameters for the newton solver.

See also newton.

Fields

verbose::Bool = false: Print some info every iteration?

step_tol::Float64 = 1e-8: If the (absolute value of) step size is smaller than this tolerance, stop.

f_tol::Float64 = step_tol: If the (absolute value of) function value is smaller than this tolerance, stop.

max_iter::Integer = 750: Maximum number of iterations before giving up.

msg = "No Newton convergence": Warning message to display if there isn't convergence within the maximum number of iterations.

l::Real = -Inf: Left bound.

r::Real = Inf: Right bound.

EquilibriumUtilities.newton — Function

Simple Newton method implementation.

See also NewtonParameters for more keyword argument options.

EquilibriumUtilities.newton — Method

newton(f_f′, x; kwargs...)

f_f′ should be a function which returns the tuple (f, f′) of function value and derivative. x is the initial guess.

EquilibriumUtilities.newton — Method

newton(f, f′, x; kwargs...)

f should be a function which returns the function value and f′ its derivative. x is the initial guess.

Small utilities

EquilibriumUtilities.EquilibriumUtilities — Module

A package of basic utility functions used commonly when computing economic equilibria.

See converge, newton, normalise!, zero_safe.

EquilibriumUtilities.normalise! — Function

normalise!(v, factor = zero_safe(first(v)))

Normalise v by some factor. Useful for normalising nominal prices (so the first element is the numeraire) or shares (with factor = sum(v)).

See also zero_safe.

EquilibriumUtilities.quietly — Method

quietly(f::Function)

Execute something without logging it (sends logs to Logging.NullLogger()).

EquilibriumUtilities.zero_safe — Method

zero_safe(x)

If x is zero, return one(x). Otherwise, return x. Useful for safely dividing by x.

EquilibriumUtilities.chunk — Method

chunk(V::AbstractVector, N::Integer)

Chunk a vector V into Base.views each of length N. Returns a Tuple of Base.views.

EquilibriumUtilities.diag_view — Method

diag_view(mat::AbstractMatrix)

Returns a view of the matrix diagonal, assuming one-based indexing.

See also offdiag_view, issquare.

EquilibriumUtilities.issquare — Method

issquare(mat::AbstractMatrix)

Is the matrix square?

See also diag_view, offdiag_view.

EquilibriumUtilities.offdiag_view — Method

offdiag_view(mat::AbstractMatrix)

Returns a view of the matrix off-diagonals, assuming one-based indexing.

See also diag_view, issquare.

EquilibriumUtilities.pretty — Function

pretty(s, var_names = keys(s); kw...)

Return a string version of s for pretty-printing.

Keywords

pad = 8: column width

digits = 4: number of digits to round to

spacer = 2: number of spaces in between columns

WrappedDict

EquilibriumUtilities.WrappedDict — Type

abstract type WrappedDict{T} <: AbstractDict{Symbol, T}

Eases creating custom structs that are basically dictionaries. (Essentially implements keys, values, length, iterate, getindex, setindex!, get, and get! by delegating to the necessary field internal_dict.)

A concrete subtype of WrappedDict must have a field internal_dict::Dict{Symbol, T}, which is the dictionary.

For an example, see this package's tests with an implementation of a Pollak demand function struct as a WrappedDict.

See also validate.

EquilibriumUtilities.validate — Method

validate(given, needed)

For each key in needed, check that it's present in given.