EquilibriumUtilities.jl Documentation

EquilibriumUtilities.EquilibriumUtilities — Module

A package of basic utility functions used commonly when computing economic equilibria.

See converge, newton, normalise!, zero_safe.

EquilibriumUtilities.ConvergeParameters — Type

ConvergeParameters{T1 <: Real, T2 <: Integer, T3}

Contains optional parameters for the converge solver.

See also converge.

Fields

diff_tol::T1 = 1e-6: If the value returned by step_diff is less than this tolerance, stop.

up_tol::T1 = zero(diff_tol): If the value returned by update is less than this tolerance, stop.

max_iter::T2 = 200: Maximum number of iterations before giving up.

msg::T3 = "No convergence": Warning message to display if there isn't convergence within the maximum number of iterations.

verbose::Bool = false: Print the diff every iteration?

EquilibriumUtilities.DampenParameters — Type

struct DampenParameters{T1 <: Real, T2 <: Integer}

Parameters for dynamicdampen!.

Fields

loosen::T1 = -0.01: by how much to loosen the dampening factor

tighten::T1 = 0.01: by how much to tighten the dampening factor

min_dampen::T1 = 0.0: the minimum value for the dampen factor

max_dampen::T1 = 0.999: the maximum value for the dampen factor

scale::T1 = 0.925: by how much to scale the reference deviation

overshooting_share::T1 = 0.5: the share for isovershooting

grace_period::T2 = 50: how long to wait between adjustments

tighten_wait::T2 = 30: how long to wait between tightenings

loosen_wait::T2 = 40: how long to wait between loosenings

EquilibriumUtilities.DampenState — Type

struct DampenState{T1 <: Real, T2 <: AbstractVector{T1}}

Fields

dampen::T1 = 0.85: the dampening factor

last_loosened::T1 = Inf: number of iterations since the dampening factor lowered

last_tightened::T1 = Inf: number of iterations since the dampening factor increased

reference_diff::T1 = Inf: the diff used when checking whether iteration is far along enough to attempt loosening again

last_deviations::T2 = Float64[] for storing deviations from the last iteration

penultimate_deviations::T2 = Float64[] for storing deviations from the penultimate iteration

history::T2 = Float64[] for storing the history of diffs

EquilibriumUtilities.NewtonParameters — Type

NewtonParameters

Contains optional parameters for the newton solver.

See also newton.

Fields

verbose::Bool = false: Print some info every iteration?

step_tol::Float64 = 1e-8: If the (absolute value of) step size is smaller than this tolerance, stop.

f_tol::Float64 = step_tol: If the (absolute value of) function value is smaller than this tolerance, stop.

max_iter::Integer = 750: Maximum number of iterations before giving up.

msg = "No Newton convergence": Warning message to display if there isn't convergence within the maximum number of iterations.

l::Real = -Inf: Left bound.

r::Real = Inf: Right bound.

EquilibriumUtilities.WrappedDict — Type

abstract type WrappedDict{T} <: AbstractDict{Symbol, T}

Eases creating custom structs that are basically dictionaries.[1] A concrete subtype of WrappedDict must have a field internal_dict::Dict{Symbol, T}, which is the dictionary.

For an example, see this package's tests with an implementation of a Pollak demand function struct as a WrappedDict.

See also validate, pretty.

EquilibriumUtilities.chunk — Method

chunk(V::AbstractVector, N::Integer)

Chunk a vector V into Base.views each of length N. Returns a Tuple of Base.views.

EquilibriumUtilities.converge — Method

converge(update::Function, step_diff::Function, init::Function; kw...) -> (converged, stalled_update)

Iterate until convergence. In particular, the problem is initiated with init(). Then, repeatedly:

apply step_diff() which should do an iteration step, then return the deviation

apply update(), which should return the size of the update

until convergence. The convergence criterion is controlled by keyword arguments; see ConvergeParameters.

The function returns a pair of booleans, the first of which indicates whether convergence was reached; the second of which indicates whether iteration was aborted because the update was too small.

For an example, see this package's tests which uses this function to solve the finite-firm CES game.

See also update!.

EquilibriumUtilities.dynamicdampen! — Method

dynamicdampen!(cs::DampenState, p::DampenParameters)

Update the dampening factor based on convergence path. Strategy:

increase by cs.tighten if cs.history isdiverging or isovershooting

make no change if any of:

within the first p.grace_period iterations

it has been under p.tighten_wait iterations since the last tightening

it has been under p.loosen_wait iterations since the last loosening

the current difference is above p.scale*cs.reference_diff

otherwise, decrease by loosen

See also DampenState, DampenParameters.

EquilibriumUtilities.isdiverging — Method

isdiverging(history)

Checks if iteration is on a bad path, by seeing if two of the past three iterations have worsened the difference. If updating is too aggressive, usually one of two things happens:

the difference blows up (if updating is too aggressive)

the difference oscillates between improving and worsening (if updating is a little too aggressive)

See also dynamicdampen!.

EquilibriumUtilities.isovershooting — Method

isovershooting(last_deviations, penultimate_deviations; share = 0.5)

Check if the convergence is overshooting each guess, as measured by a share of the deviations flipping signs between the last two iterations. If dampening isn't strong enough, the deviations alternate between positive and negative, wasting time by bouncing back and forth. Increasing the dampening can drastically speed up convergence by preventing this oscillation.

See also dynamicdampen!.

EquilibriumUtilities.newton — Function

Simple Newton method implementation.

See also NewtonParameters for more keyword argument options.

EquilibriumUtilities.newton — Method

newton(f_f′, x; kwargs...)

f_f′ should be a function which returns the tuple (f, f′) of function value and derivative. x is the initial guess.

EquilibriumUtilities.newton — Method

newton(f, f′, x; kwargs...)

f should be a function which returns the function value and f′ its derivative. x is the initial guess.

EquilibriumUtilities.normalise! — Function

normalise!(v, factor = zero_safe(first(v)))

Normalise v by some factor. Useful for normalising nominal prices (so the first element is the numeraire) or shares (with factor = sum(v)).

See also zero_safe.

EquilibriumUtilities.pretty — Function

pretty(s::WrappedDict, var_names = keys(s); kw...)

Return a string version of s for pretty-printing.

Keywords

pad = 8: column width

digits = 4: number of digits to round to

spacer = 2: number of spaces in between columns

EquilibriumUtilities.update! — Method

update!(main, secondary; norm = infnorm_pctdev, dampen = 0.5)

Update main according to secondary, with a dampening factor. Useful for iterative algorithms. Once complete, main will hold the updated value and secondary will hold main's original value (to keep a record of previous iteration). Returns the norm of the update step (useful for breaking iteration if the step size is too small).

EquilibriumUtilities.validate — Method

validate(given, needed)

For each key in needed, check that it's present in given.

EquilibriumUtilities.zero_safe — Method

zero_safe(x)

If x is zero, return one(x). Otherwise, return x. Useful for safely dividing by x.

1Essentially implements keys, values, length, iterate, getindex, setindex!, get, and get! by delegating to the necessary field internal_dict.