EquilibriumUtilities.jl Documentation

EquilibriumUtilities.EquilibriumUtilities — Module

A package of basic utility functions used commonly when computing economic equilibria.

See converge, normalise!, zero_safe, newton.

EquilibriumUtilities.WrappedDict — Type

abstract type WrappedDict{T} <: AbstractDict{Symbol, T}

Eases creating custom structs that are basically dictionaries.[1] A concrete subtype of WrappedDict must have a field internal_dict::Dict{Symbol, T}, which is the dictionary.

For an example, see this package's tests with an implementation of a Pollak demand function struct as a WrappedDict.

See also validate, pretty.

EquilibriumUtilities.converge — Method

converge(update::Function, step_diff::Function, init::Function; kwargs...)

Iterate until convergence. In particular, the problem is initiated with init(). Then, repeatedly apply step_diff() which should do an iteration step, then return the diff. converge will compare the returned difference with tol; if it's smaller, then converge is reached and iteration stops. Otherwise, it will apply update(), which should return the size of the update.

The function returns a pair of booleans, the first of which signals whether convergence was reached; the second signals whether iteration was aborted because the update was too small.

For an example, see this package's tests which uses this function to solve the finite-firm CES game.

See also update!, dampen, v_diff.

Keywords

diff_tol::Real = 1e-6 : If the value returned by step_diff is less than this tolerance, stop.

up_tol::Real = zero(diff_tol) : If the value returned by update is less than this tolerance, stop.

max_iter::Integer = 200 : Maximum number of iterations before giving up.

msg = "No convergence" : Warning message to display if there isn't convergence within the maximum number of iterations.

history = nothing : Container in which to store the iteration history of diffs. Useful to check speed of convergence. If nothing is provided, saves no history.

verbose::Bool = false : Print the diff every iteration?

EquilibriumUtilities.dampen — Method

dampen(history; kwargs...)

Given an iteration history, return a dampening factor. At the moment, just uses the last value of history.

Keywords

slow = 0.95 : the dampening factor for slow updating (when diff is above 10).

med = 0.75 : the dampening factor for medium updating (when diff is above 1).

fast = 0.5 : the dampening factor for fast updating.

EquilibriumUtilities.dynamic_dampen — Method

dynamic_dampen(dampen, last_loosened, last_tightened, reference_diff, last_devs, penultimate_devs, history; kw...)

Update the dampening factor based on convergence path. Strategy:

increase by tighten if history isdiverging or isovershooting

make no change if any of:

* within the first `grace_period` iterations
* it has been under `tighten_wait` iterations since the last tightening
* it has been under `loosen_wait` iterations since the last loosening
* the current difference is above `scale*reference_diff`
* the `tail` of `history` [`!isconvex`](@ref), with tolerance `convex_tol`

otherwise, decrease by loosen

EquilibriumUtilities.isconvex — Method

isconvex(history; tail = 25, tol = 0.005)

Check if the last tail entries in history is convex. Useful for dynamically updating the dampening factory.

See also dynamic_dampen.

EquilibriumUtilities.isdiverging — Method

isdiverging(history)

Checks if iteration is on a bad path, but seeing if two of the past 3 iterations have worsened the difference. Useful for dynamically updating the dampening factor.[1]

See also dynamic_dampen.

[^1]: If updating is too aggressive, usually one of two things happens. Either the difference blows up (if updating is _too_ aggressive) or the difference oscillates between improving and worsening (if updating is a little too aggressive).

EquilibriumUtilities.isovershooting — Method

isovershooting(last_dev, penultimate_dev; share = 0.5)

Check if the convergence is overshooting each guess, as measured by a share of the deviations flipping signs between the last two iterations. Useful for dynamically updating the dampening factor.[1]

See also dynamic_dampen.

[^1]: If dampening isn't strong enough, the devations alternate between positive and negative. Convergence wastes times bouncing back and forth, usually slowing down. Increasing the dampening can drastically speed up convergence by preventing this oscillation.

EquilibriumUtilities.newton — Function

Simple Newton method implementation.

Keywords

step_tol = 1e-8 : If the (absolute value of) step size is smaller than this tolerance, stop.

f_tol = step_tol : If the (absolute value of) function value is smaller than this tolerance, stop.

max_iter = 750 : Maximum number of iterations before giving up.

l = -Inf : Left bound.

r = Inf : Right bound.

msg = "No Newton convergence" : Warning message to display if there isn't convergence within the maximum number of iterations.

verbose::Bool = false : Print some info every iteration?

EquilibriumUtilities.newton — Method

newton(f_f′, x; kwargs...)

f_f′ should be a function which returns the tuple (f, f′) of function value and derivative. x is the initial guess.

EquilibriumUtilities.newton — Method

newton(f, f′, x; kwargs...)

f should be a function which returns the function value and f′ its derivative. x is the initial guess.

EquilibriumUtilities.normalise! — Function

normalise!(v, factor = zero_safe(first(v)))

Normalise v by some factor. Useful for normalising nominal prices (so the first element is the numeraire) or shares (with factor = sum(v)).

See also zero_safe.

EquilibriumUtilities.pretty — Function

pretty(s::WrappedDict, var_names = keys(s); kw...)

Return a string version of s for pretty-printing.

Keywords

pad = 8: column width

digits = 4: number of digits to round to

spacer = 2: number of spaces in between columns

EquilibriumUtilities.update! — Method

update!(main, secondary; dampen = 0.75, v_diff = v_diff)

Update main according to secondary, with a dampening factor. Useful for iterative algorithms. Once complete, main will hold the updated value and secondary will hold main's original value (to keep a record of previous iteration). Returns the v_diff of the update step (useful for breaking iteration if the step size is too small).

EquilibriumUtilities.v_diff — Method

v_diff(v1, v2)

Calculate the distance between two vectors as the sum of element-wise absolute difference.

EquilibriumUtilities.validate — Method

validate(given, needed)

For each key in needed, check that it's present in given.

EquilibriumUtilities.zero_safe — Method

zero_safe(x)

If x is zero, return one(x). Otherwise, return x. Useful for safely dividing by x.

1Essentially implements keys, values, length, iterate, getindex, setindex!, get, and get! by delegating to the necessary field internal_dict.